added the header stddef

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -7,15 +7,14 @@
  */
 size_t print_list(const list_t *h)
 {
-	int count = 0;
-
+	size_t count = 0;
 
 	while (h != NULL)
 	{
 		if (h->str == NULL)
 			printf("[0] (nill)\n");
 		else
-			printf("[%i] %s\n", h->len, h->str);
+			printf("[%d] %s\n", h->len, h->str);
 
 		h = h->next;
 		count++;
